refactor(rendering): :recycle: move Color from utils to gfx
